backend/requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.7
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
flower==2.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==6.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-telegram-bot==22.3
pytz==2025.2
redis==5.0.1
requests==2.31.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.5.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
